requirements.txt
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
filelock==3.18.0
flake8==7.3.0
identify==2.6.12
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre-commit==4.2.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyYAML==6.0.2
tomli==2.2.1
typing-extensions==4.14.1
virtualenv==20.31.2
